renamed schemas to password service to better reflect what it does
