+++ apps/backend/app/api/auth.py
@@ -9,7 +9,7 @@ from fastapi import (
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from services.auth.login import LoginRequest, LoginResponse
-from services.auth.schemas import verify_password
+from auth.password_service import verify_password
 from services.auth.signup import SignupRequest, SignupResponse
 from services.database_manager.database import get_db
 from services.database_manager.managers.user_manager import user_manager
