+++ services/auth/login.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel, EmailStr, field_validator
 
-from services.auth.schemas import validate_password_strength
+from auth.password_service import validate_password_strength
 
 
 class LoginRequest(BaseModel):
